downstream/segmentation/requirements.txt
torch==2.0.0
torchvision
mmengine>=0.5.0,<1.0.0
albumentations>=0.3.2 --no-binary qudida,albumentations
ftfy
regex
timm
matplotlib
numpy
packaging
prettytable
scipy
mmsegmentation==1.2.2
tensorboard
future
setuptools==69.5.1
